+++ examples/minimize.cpp
@@ -21,6 +21,7 @@
 using namespace std;
 using namespace TCLAP;
 
+//!What, after all, *is* the volume of a d-dimensional sphere?
 scalar sphereVolume(scalar radius, int dimension)
     {
     if(dimension == 1)
@@ -33,9 +34,7 @@ scalar sphereVolume(scalar radius, int dimension)
     };
 
 /*!
-command line parameters help identify a data directory and a filename... the output is a text file
-(in the data/ directory rooted here) containing easy-to-read fourier transforms of the height-map
-representation of the extremal interfaces for each point in time
+testing the minimizers
 */
 int main(int argc, char*argv[])
 {
@@ -182,7 +181,7 @@ int main(int argc, char*argv[])
         p = trace(Pressure) / (1.0*DIMENSION);
         printf("p=%g\n\n",p);
         };
-    
+
 
 
     cudaProfilerStart();